Added ability for units to move and attack
Created a UnitTrigger component which simply responds to game state of whether a unit is selected or not.
It then generates the movement and combat area. Later it will be responsible for update the UI with respect to unit-specific capabilities such as move, actions, attack, etc...
Its basically part of the unit logic decoupled into a component

Minor updates to DomGametile to better reflect the new logic.
Updated player colouring logic; player 'numbers' are assigned on game creation.
Made world generation step not require user input. World will be generated and game started immediately