Changed underlying world representation of units
Previously Units were placed directly onto the layer with the structure { <coordinate>:<unit>, ... }
Now the unit-layer only contains the unitId which must then be looked-up later for the actual unit if required.
This has the benefit of allowing actions to refer to units by unitId instead of an otherwise unintuitive and cumbersome coordinate. It makes the player-interface a bit easier to work with but forces the store:world to manage the coordination