More visual changes for units based on energy
If a unit has moved, we change the opacity slightly.
If a unit has actioned, we reduce the opacity even further.
This is to make it more obvious which units are still capable of being interacted with or not.
- UI/UX might change later
When a player surrenders (i.e. relinquishes the game), all their units have their energy eliminated, effectively making it much clearer who is the winner