Refactored combat logic
Since we've added some states for combat (before, during, after) we need to rework some of the logic to handle this transition seamlessly.
This required the addition of a 'CombatTrigger' and a new store 'Combat'
The trigger works similarly to PlayerPhaseTrigger and is simply responsible for state transitions - Later the two should be combined so we have a single 'game state engine'
The 'Combat' store is where all combat logic resides.
- Upon starting, the attacker and defender coordinates (source / destination) are required. Everything else is calculated based on the current world state.
- Combat occurs using the internal information
- Clean-up
- Then transition back to PLAYER_TURN
In theory, the player's controls might be disabled/restricted whilst in combat.
Using the transitions also allows us to do animations